refactor: move execution utils to execution folder
the Publisher and Bundler abstractions are such core parts of the execution algorithm that they probably belong within the execution folder (following the pattern of the  map/flatten asyncIterable helpers)
