+++ src/execution/bundler.ts
@@ -1,4 +1,4 @@
-import type { Maybe } from './Maybe';
+import type { Maybe } from '../jsutils/Maybe';
 
 export interface IBundler<TDataResult, TErrorResult> {
   queueData: (index: number, result: TDataResult) => void;
